+++ example/main.dart
@@ -3,8 +3,8 @@ import 'package:angel_hot/angel_hot.dart';
 import 'server.dart';
 
 main() async {
-  var hot = new HotReloader(createServer, [
-    new Directory('src'),
+  var hot = HotReloader(createServer, [
+    Directory('src'),
     'server.dart',
     // Also allowed: Platform.script,
     Uri.parse('package:angel_hot/angel_hot.dart')